+++ src/errors/StructureError.ts
@@ -1,6 +1,7 @@
+import { AnalysisError } from './AnalysisError'
 import { STRUCTURE_ERROR_UNCAUGHT } from './codes'
 
-export class StructureError extends Error {
+export class StructureError extends AnalysisError {
   public readonly code: typeof STRUCTURE_ERROR_UNCAUGHT
 
   constructor(message: string) {