+++ src/errors/codes.ts
@@ -2,3 +2,4 @@ export const GENERIC_FAILURE = -1
 export const STRUCTURE_ERROR_UNCAUGHT = 2
 export const SOURCE_MISSING_ERROR = 3
 export const SOURCE_PARSE_ERROR = 4
+export const WRONG_ANALYZER_FOR_METHOD = 5